🐛 fix(workspace): move slug format rules into @lobechat/const
#19015 put isWorkspaceSlugFormatValid in @lobechat/business-const and had the
business-server workspace stub import it. business-const is a business slot
that deployments override with their own package, so a symbol only the OSS
copy exports breaks every downstream type-check the moment the stub uses it.

The format rules are generic (they define what a slug is), so they now live in
@lobechat/const next to the other workspace constants; the stub imports from
there and business-const drops its copy.
