+++ packages/business-server/package.json
@@ -7,7 +7,7 @@
   },
   "dependencies": {
     "@lobechat/agent-gateway-client": "workspace:*",
-    "@lobechat/business-const": "workspace:*",
+    "@lobechat/const": "workspace:*",
     "@lobechat/database": "workspace:*",
     "@lobechat/model-runtime": "workspace:*",
     "@lobechat/trpc": "workspace:*",
