+++ packages/business/const/src/index.ts
@@ -3,7 +3,6 @@ import { BRANDING_PROVIDER } from './branding';
 export * from './branding';
 export * from './llm';
 export * from './url';
-export * from './workspaceSlug';
 
 export const ENABLE_BUSINESS_FEATURES = false;
